refactor: create services instead of helpers
